Move update-monodoc trigger from monodoc-base (in mono)
The commands being executed from the trigger are in monodoc-browser, so
it makes sense for the trigger to be here too, then we don't have have
base attempt to check if -browser is configured, which we could never
get right.

We need to have Replaces: and Breaks: monodoc-base (<< 2.10.8.1-3) for
the moved update-monodoc. -3 is the version which will stop shipping
update-monodoc; we need to make sure that users don't have this on their
system after the trigger has moved.

LP: #972751